Time: 3 ms (97.83%) | Memory: 20.1 MB (61.66%) - LeetSync

+++ 98-validate-binary-search-tree/validate-binary-search-tree.cpp
@@ -12,7 +12,7 @@
 class Solution {
 public:
     ////gambling algo//////
-    long val=rand(),n=val;
+    int val=rand(),n=val;
     bool flag=true,flag1=true;
     void fn(TreeNode* root)
     {
